// petforest-fe-cms/src/utils/enums/setting/setting-mngt.enums.ts
export enum SETTING_TABS {
  SERVICE = 'SERVICE',
  VARIETIES = 'VARIETIES',
  CORRESPONDENT = 'CORRESPONDENT',
  PROGRESS = 'PROGRESS',
  PARTNER = 'PARTNER',
  WAITING_ROOM = 'WAITING_ROOM',
  HOSPITAL_SETTLE_METHOD = 'HOSPITAL_SETTLE_METHOD',
}

export enum SETTING_DIVISION {
  GENERAL = 'general',
  LUCETTE = 'lucette',
}

export enum SETTING_SERVICE {
  STANDARD = 'standard',
  FOREST = 'forest',
  PREMIUM = 'premium',
  EXCLUSIVE = 'exclusive',
}

export enum SETTING_PET_TYPE {
  DOG = 'dog',
  CAT = 'cat',
  BIRD = 'bird',
  HAMSTER = 'hamster',
  ETC = 'etc',
}