// petforest-fe-cms/src/utils/enums/side-menu-type.ts
export enum MenuKeys {
  //Home
  HOME = 'HOME',

  //Dashboard
  DASHBOARD = 'DASHBOARD',

  //Funeral MNGT
  FUNERAL = 'FUNERAL',
  FUNERAL_RESERVATION = 'FUNERAL_RESERVATION',
  FUNERAL_PNT_SETTING = 'FUNERAL_PNT_SETTING',
  FUNERAL_LUCETTE = 'FUNERAL_LUCETTE',
  FUNERAL_STOCK = 'FUNERAL_STOCK',
  FUNERAL_NOTI_HISTORY = 'FUNERAL_NOTI_HISTORY',

  //Settlement
  SETTLEMENT = 'SETTLEMENT',
  SETTLEMENT_ADJUSTMENT = 'SETTLEMENT_ADJUSTMENT',
  SETTLEMENT_COUPON = 'SETTLEMENT_COUPON',

  //Ossuary
  OSSUARY = 'OSSUARY',

  //Material
  MATERIAL = 'MATERIAL',
  MATERIAL_FUNERAL_SUPPLIES = 'MATERIAL_FUNERAL_SUPPLIES',
  MATERIAL_LUCETTE_SUPPLIES = 'MATERIAL_LUCETTE_SUPPLIES',
  MATERIAL_OTHER_FUNERAL_SUPPLIES = 'MATERIAL_OTHER_FUNERAL_SUPPLIES',
  MATERIAL_EXPENDABLES_SUPPLIES = 'MATERIAL_EXPENDABLES_SUPPLIES',

  //Facility
  FACILITY = 'FACILITY',

  //Collaborating Hospital
  COLLABORATING = 'COLLABORATING',
  COLLABORATING_SALES_CALENDAR = 'COLLABORATING_SALES_CALENDAR',
  COLLABORATING_PARTNER_HOSPITAL = 'COLLABORATING_PARTNER_HOSPITAL',
  COLLABORATING_SALE_LIST = 'COLLABORATING_SALE_LIST',
  COLLABORATING_DELIVERY_PROMOTIONAL_MATERIAL = 'COLLABORATING_DELIVERY_PROMOTIONAL_MATERIAL',
  COLLABORATING_SETTLEMENT = 'COLLABORATING_SETTLEMENT',

  //Post MNGT
  POST = 'POST',

  //Statistic
  STATISTIC = 'STATISTIC',
  STATISTIC_BASIC = 'STATISTIC_BASIC',
  STATISTIC_REGION = 'STATISTIC_REGION',
  STATISTIC_ARTICLE = 'STATISTIC_ARTICLE',

  //Manager
  MANAGER = 'MANAGER',
  MANAGER_RESERVATION_TIME = 'MANAGER_RESERVATION_TIME',
  MANAGER_USER = 'MANAGER_USER',
  MANAGER_BRANCH = 'MANAGER_BRANCH',

  //Setting
  SETTING = 'SETTING',

  RECEIVE_RESERVATION = 'RECEIVE_RESERVATION',
}